requirements-dev.txt
pytest==8.3.4
pytest-asyncio>=0.23
requests>=2.28
python-telegram-bot>=20.0
httpx>=0.27
python-dotenv>=1.0
pydantic-settings>=2.0
